feat: remove edit offer screen
