+++ frontend/app/oferta/[ofertaid].tsx
@@ -1,10 +1,10 @@
-import { Text, View, ActivityIndicator, StyleSheet, TouchableOpacity, Image, Platform, ScrollView, Linking, Alert, Modal } from "react-native";
+import { Text, View, ActivityIndicator, StyleSheet, TouchableOpacity, Image, Platform, ScrollView, Alert, Modal } from "react-native";
+import CheckBox from '@react-native-community/checkbox';
 import React, { useState, useEffect } from "react";
-import { router, useLocalSearchParams, useRouter } from "expo-router";
+import { useLocalSearchParams, useRouter } from "expo-router";
 import { FontAwesome5, MaterialIcons, Entypo } from "@expo/vector-icons";
 import colors from "frontend/assets/styles/colors";
 import { useAuth } from "@/contexts/AuthContext";
-import { Ionicons } from '@expo/vector-icons';
 import MaterialCommunityIcons from '@expo/vector-icons/MaterialCommunityIcons';
 import defaultCompanyLogo from "frontend/assets/images/defaultCompImg.png"
 import defaultCamImage from "../../assets/images/defaultAvatar.png";
@@ -177,11 +177,26 @@ export default function OfertaDetalleScreen() {
 
     const handleLoginRedirect = () => {
         router.push("/login")
-    };    
+    };
 
-    const handleEditarOferta = () => {
-        router.push(`/oferta/editar/${ofertaid}`);
+    const handleDeleteOffer = async () => {
+    try {
+        const response = await fetch(`${BACKEND_URL}/ofertas/${ofertaid}`, {
+        method: "DELETE",
+        headers: {
+            'Authorization': `Bearer ${userToken}`
+        }
+        });
+
+        if (response.ok) {
+        router.replace("/miperfil");    
+        } else {
+            Alert.alert("Error", "No se pudo eliminar la oferta.");
+        }
+    } catch (error) {
+        console.error("Error al eliminar la oferta:", error);
     }
+    };
 
     const renderOfferCard = () => {
         return (
@@ -223,10 +238,6 @@ export default function OfertaDetalleScreen() {
                                     </TouchableOpacity>
                                 )
                             ) : <></>
-                        ) : user.rol === 'EMPRESA' && user.id === offerData.empresa.id ? (
-                                <TouchableOpacity style={styles.solicitarButton} onPress={handleEditarOferta}>
-                                    <Text style={styles.solicitarButtonText}>Editar Oferta</Text>
-                                </TouchableOpacity>
                         ) : <></>
                     ) : (
                         <TouchableOpacity style={styles.solicitarButton} onPress={handleLoginRedirect}>
@@ -409,6 +420,15 @@ export default function OfertaDetalleScreen() {
                                             </View>
                                         </View>
                                     ))}
+                                </View>
+                                <View style={styles.separator}/>
+                                <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+                                    <TouchableOpacity
+                                        style={styles.deleteButton}
+                                        onPress={handleDeleteOffer}
+                                    >
+                                        <Text style={styles.deleteButtonText}>Eliminar Oferta</Text>
+                                    </TouchableOpacity>
                                 </View>
                             </>
                         ) : (
@@ -656,4 +676,17 @@ const styles = StyleSheet.create    ({
         marginLeft: 10,
         marginTop: Platform.OS === 'ios' ? 30 : 10,
     },
+    deleteButton: {
+        flex: 1,
+        paddingVertical: 12,
+        marginRight: 10,
+        borderRadius: 12,
+        alignItems: "center",
+        backgroundColor: "#D14F45",
+    },
+    deleteButtonText: {
+        color: "white",
+        fontSize: 16,
+        fontWeight: "bold",
+    }
 });
